+++ src/app/logs/logs-modal/logs-modal.component.html
@@ -25,24 +25,6 @@
 
 <ng-template #all>
     <div [hidden]="tabActive!=='All'">
-
-<!--        <div class="grid grid-cols-12 auto-rows-auto bg-slate-50 border mb-5 py-1 px-1 gap-4">-->
-<!--            <div class="col-span-6">-->
-<!--                @for(log of logsTree.before[data.logIndex]; track $index){-->
-<!--                    @if(log){-->
-<!--                        <app-log-data [data]="log"></app-log-data>-->
-<!--                    }-->
-<!--                }-->
-<!--            </div>-->
-<!--            <div class="col-span-6">-->
-<!--                @for(log of logsTree.after[data.logIndex]; track $index){-->
-<!--                    @if(log){-->
-<!--                        <app-log-data [data]="log"></app-log-data>-->
-<!--                    }-->
-<!--                }-->
-<!--            </div>-->
-<!--        </div>-->
-
         <div class="grid grid-cols-2 auto-rows-auto bg-slate-50 border mb-5 py-1 px-1 gap-x-4">
             @for(log of logsTree.before[data.logIndex]; track $index){
                 @if(log){
