+++ src/app/logs/logs-modal/logs-modal.component.ts
@@ -1,11 +1,8 @@
-import {Component, Input, signal, ViewChild} from '@angular/core';
-import {FormBuilder, FormsModule, ReactiveFormsModule, Validators, FormGroup, FormGroupName,} from "@angular/forms";
-import {JsonPipe, KeyValuePipe, NgClass, NgIf, NgStyle, NgTemplateOutlet, UpperCasePipe} from "@angular/common";
+import {Component, Input} from '@angular/core';
+import {FormsModule, ReactiveFormsModule} from "@angular/forms";
+import {NgTemplateOutlet} from "@angular/common";
 import {ModalService} from "../../components/modal/modal.service";
-import {ModalTypes} from "../../components/modal/modal-types";
-import {CellColor} from "../../components/wi-table/wi-table.component";
 import {LogDataComponent} from "../log-data/log-data.component";
-import {log} from "@angular-devkit/build-angular/src/builders/ssr-dev-server";
 
 @Component({
   selector: 'app-logs-modal',
@@ -15,8 +12,6 @@ import {log} from "@angular-devkit/build-angular/src/builders/ssr-dev-server";
     ReactiveFormsModule,
     NgTemplateOutlet,
     LogDataComponent,
-    JsonPipe,
-    NgClass,
   ],
   templateUrl: './logs-modal.component.html',
   styleUrl: './logs-modal.component.scss'
@@ -30,8 +25,6 @@ export class LogsModalComponent {
   };
   @Input() data: any;
   public tabActive = 'All';
-  currentID = signal(null);
-  oldParentID = null; // need for back if we wanna change parent
   isParentChosen: any = false;
 
   constructor(
@@ -46,6 +39,5 @@ export class LogsModalComponent {
     if (!this.logsTree) {
       return
     }
-    console.log(this.logsTree);
   }
 }
