+++ src/app/utils/utils.ts
@@ -1,3 +1,8 @@
+export enum DiffColor {
+    delete = '#FFC7C2', // red
+    update = '#F4F3CD', // yellow
+    create = '#CDF4D3', // green
+}
 // Функція, яка перетворює Enum в {id:number, name:string}[]
 export function dropDownComponentListFromEnum(enumData: any): Array<{ id: number; name: string }> {
     return Object.values(enumData).map((value, index) => ({
